Clip alignment to a subgraph
