+++ src/alignment/tests.rs
@@ -374,224 +374,3 @@ fn alignment_real_gzipped() {
 
 //-----------------------------------------------------------------------------
 
-// Tests for `Difference`.
-
-fn check_difference(seq: &[u8], truth: &[Difference], name: &str, normalize: bool) {
-    let result = Difference::parse(seq);
-    assert!(result.is_ok(), "Failed to parse the difference string for {}: {}", name, result.err().unwrap());
-    let mut result = result.unwrap();
-    if normalize {
-        result = Difference::normalize(result);
-    }
-    assert_eq!(result.len(), truth.len(), "Wrong number of operations for {}", name);
-    for i in 0..truth.len() {
-        assert_eq!(result[i], truth[i], "Wrong operation at position {} for {}", i, name);
-    }
-}
-
-fn check_to_bytes(ops: &[Difference], truth: &[u8], target_sequence: &[u8], name: &str) {
-    let result = Difference::to_bytes(ops, target_sequence);
-    assert_eq!(result, truth, "Wrong byte representation for {}", name);
-}
-
-#[test]
-fn difference_empty() {
-    let name = "empty";
-    let seq = b"";
-    let truth = [];
-    check_difference(seq, &truth, name, false);
-    let target_sequence = b"";
-    check_to_bytes(&truth, seq, target_sequence, name);
-}
-
-#[test]
-fn difference_single() {
-    {
-        let name = "matching sequence";
-        let seq = b"=ACGT";
-        let truth = [ Difference::Match(4) ];
-        check_difference(seq, &truth, name, false);
-        let expected = b":4"; // Converted to match length.
-        let target_sequence = b"ACGT";
-        check_to_bytes(&truth, expected, target_sequence, name);
-    }
-    {
-        let name = "match length";
-        let seq = b":5";
-        let truth = [ Difference::Match(5) ];
-        check_difference(seq, &truth, name, false);
-        let target_sequence = b"GATTA";
-        check_to_bytes(&truth, seq, target_sequence, name);
-    }
-    {
-        let name = "mismatch";
-        let seq = b"*AC";
-        let truth = [ Difference::Mismatch(b'C') ];
-        check_difference(seq, &truth, name, false);
-        let target_sequence = b"A";
-        check_to_bytes(&truth, seq, target_sequence, name);
-    }
-    {
-        let name = "insertion";
-        let seq = b"+ACGT";
-        let truth = [ Difference::Insertion(b"ACGT".to_vec()) ];
-        check_difference(seq, &truth, name, false);
-        let target_sequence = b"";
-        check_to_bytes(&truth, seq, target_sequence, name);
-    }
-    {
-        let name = "deletion";
-        let seq = b"-ACGT";
-        let truth = [ Difference::Deletion(4) ];
-        check_difference(seq, &truth, name, false);
-        let target_sequence = b"ACGT";
-        check_to_bytes(&truth, seq, target_sequence, name);
-    }
-}
-
-#[test]
-fn difference_multi() {
-    // All of these are from real reads mapped with Giraffe.
-    // Target sequences only contain the relevant bases.
-    {
-        let name = "ERR3239454.129477191 fragment 2";
-        let seq = b":24*CG:78-C:35*TG:2*TG:1*TG:6";
-        let truth = [
-            Difference::Match(24),
-            Difference::Mismatch(b'G'),
-            Difference::Match(78),
-            Difference::Deletion(1),
-            Difference::Match(35),
-            Difference::Mismatch(b'G'),
-            Difference::Match(2),
-            Difference::Mismatch(b'G'),
-            Difference::Match(1),
-            Difference::Mismatch(b'G'),
-            Difference::Match(6)
-        ];
-        check_difference(seq, &truth, name, false);
-        let mut target_sequence = vec![b'-'; 151];
-        target_sequence[24] = b'C';
-        target_sequence[103] = b'C';
-        target_sequence[139] = b'T';
-        target_sequence[142] = b'T';
-        target_sequence[144] = b'T';
-        check_to_bytes(&truth, seq, &target_sequence, name);
-    }
-    {
-        let name = "ERR3239454.11251898 fragment 1";
-        let seq = b":129+TTGAGGGGGTATAAGAGGTCG";
-        let truth = [
-            Difference::Match(129),
-            Difference::Insertion(b"TTGAGGGGGTATAAGAGGTCG".to_vec())
-        ];
-        check_difference(seq, &truth, name, false);
-        let target_sequence = vec![b'-', 129];
-        check_to_bytes(&truth, seq, &target_sequence, name);
-    }
-    {
-        let name = "ERR3239454.97848632 fragment 1";
-        let seq = b":111*GT:20*CA:6*GT:2*GT:3*GT:3";
-        let truth = [
-            Difference::Match(111),
-            Difference::Mismatch(b'T'),
-            Difference::Match(20),
-            Difference::Mismatch(b'A'),
-            Difference::Match(6),
-            Difference::Mismatch(b'T'),
-            Difference::Match(2),
-            Difference::Mismatch(b'T'),
-            Difference::Match(3),
-            Difference::Mismatch(b'T'),
-            Difference::Match(3)
-        ];
-        check_difference(seq, &truth, name, false);
-        let mut target_sequence = vec![b'-'; 150];
-        target_sequence[111] = b'G';
-        target_sequence[132] = b'C';
-        target_sequence[139] = b'G';
-        target_sequence[142] = b'G';
-        target_sequence[146] = b'G';
-        check_to_bytes(&truth, seq, &target_sequence, name);
-    }
-}
-
-#[test]
-fn difference_case() {
-    {
-        // Lower case mismatch.
-        let seq = b"*ac";
-        let truth = [ Difference::Mismatch(b'C') ];
-        check_difference(seq, &truth, "lower case mismatch", false);
-    }
-    {
-        // Lower case insertion.
-        let seq = b"+acgt";
-        let truth = [ Difference::Insertion(b"ACGT".to_vec()) ];
-        check_difference(seq, &truth, "lower case insertion", false);
-    }
-}
-
-fn invalid_difference(seq: &[u8], name: &str) {
-    let result = Difference::parse(seq);
-    assert!(result.is_err(), "Parsed an invalid difference string: {}", name);
-}
-
-#[test]
-fn difference_invalid() {
-    invalid_difference(b"ABC", "invalid operation at start");
-    invalid_difference(b"+AC:", "empty operation at end");
-    invalid_difference(b"~AC30AC", "intron / splice operation");
-    invalid_difference(b":123A", "match length is not an integer");
-    invalid_difference(b"+GATTACA:", "empty match length");
-    invalid_difference(b"*ACT", "mismatch is too long");
-    invalid_difference(b"*A", "mismatch is too short");
-}
-
-#[test]
-fn difference_len() {
-    {
-        let diff = Difference::Match(42);
-        assert_eq!(diff.len(), 42, "Wrong length for a match");
-    }
-    {
-        let diff = Difference::Mismatch(b'A');
-        assert_eq!(diff.len(), 1, "Wrong length for a mismatch");
-    }
-    {
-        let diff = Difference::Insertion(b"GATTACA".to_vec());
-        assert_eq!(diff.len(), 7, "Wrong length for an insertion");
-    }
-    {
-        let diff = Difference::Deletion(42);
-        assert_eq!(diff.len(), 42, "Wrong length for a deletion");
-    }
-}
-
-#[test]
-fn difference_normalize() {
-    {
-        let seq = b"=ACGT:4*AC*GT:2:3=ACT*AC-ACGT-ACGT+GATTACA+CAT";
-        let truth = [
-            Difference::Match(8),
-            Difference::Mismatch(b'C'),
-            Difference::Mismatch(b'T'),
-            Difference::Match(8),
-            Difference::Mismatch(b'C'),
-            Difference::Deletion(8),
-            Difference::Insertion(b"GATTACACAT".to_vec()),
-        ];
-        check_difference(seq, &truth, "normalized", true);
-    }
-    {
-        let seq = b"=:0+-*AC:30=GATTA+-:0=";
-        let truth = [
-            Difference::Mismatch(b'C'),
-            Difference::Match(35),
-        ];
-        check_difference(seq, &truth, "normalized with empty operations", true);
-    }
-}
-
-//-----------------------------------------------------------------------------
-
